% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybincon.R
\name{mybincon}
\alias{mybincon}
\title{mybincon}
\usage{
mybincon(iter, n, p, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number in each sample}

\item{p}{probability of success}

\item{...}{pass additional options to the argument}
}
\value{
a summary
}
\description{
Produces a simulation for the binomial distribution with a bar-plot and comparison to the true proportion values
}
\examples{
\dontrun{mybincon(1000, 10 , 0.6)}
}
