% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MATH4753GLab6.R
\name{summary.MATH4753GLab6}
\alias{summary.MATH4753GLab6}
\title{summary.MATH4753GLab6}
\usage{
\method{summary}{MATH4753GLab6}(object, ...)
}
\arguments{
\item{object}{a numerical object}

\item{...}{pass additional options to the argument}
}
\value{
a vector of sample sd's of length "iter"
}
\description{
Summarizes the binomial simulation by returning a vector of sample standard deviations
}
\examples{
\dontrun{summary.MATH4753GLab6)(x = mpg)}
}
